Scan for stocks that close higher dialy
Add a scanner to look for stocks that are closing higher every day. The user needs to modify the beginning date if necessary. The scanner will alert if the date mentioned is not included in the data. The user also needs to provide the path to the data in CSV files. Each CSV file should pertain to one particular stock.

The algorithm checks for the the condition when the open of the current day candle i more than the close of the previous day candle. The open of the current day candle can be a little bit lower than the close of the previous day candle also. Currently this is hardcoded to 98% of the previous day candle. Both the candles need to be green. And the law of the current day candle should not be less then the open of the previous day candle.

The output includes the list of stocks and their open and close with satisfy this criteria

unit test: checked it on the list of stocks on nifty 500 for some dates in April 2018

Co-Authored-By: aprataksh <aprataksh@users.noreply.github.com>