add the ability to append newly downloaded data to historical data; more documentation
add the ability to append a newly downloaded data to the historical data.

One problem is that the newly downloaded data has a header which also gets appended to the historical data. Possible ways to resolve this included reading the files ussing a pandas data. frame or using a CSV reader and writer. in the former case, the code ends up depending the row number to every row. In the latter case, the writer ends up adding a blank row.

So, for now, we will live with the multiple headers being a part of historical data.

Also added comments in the global configuration settings.

Unit test: checked and validated that the data was being appended to a small subset of NSE 500 stock data..

Also downloaded fresh new data for different intervals..