add the ability to calculate resistance
at the ability to calculate the resistance by looking at the price range in which a particular security has spent the maximum amount of time. This is achieved by first determining the interval of price of a particular security by looking at its entire price range during the given period for which data is provided.

Next, for each such price interval, the frequency of time periods is calculated. The price interval which has got the maximum frequency of time periods is provided as resistance.

The entire data including the price intervals and the frequency is dumped as output in a CSV file.

Co-Authored-By: aprataksh <aprataksh@users.noreply.github.com>