updated task pyalgotrade
while implementing the MA CD divergences strategy, it was discovered that the pyalgotrade does not provide proper data after calculating the signal line. This appears to be a bug and we can contribute to the open source software.

Also need to create a common library of utilities and the path to keep the library