alert the user to forthcoming dates of weekly and monthly results
alert the user to forthcoming dates of weekly and monthly results. If the user provides a list of symbols, it will also tell if any of the forthcoming results pertain to the companies whose symbols have been included by the user in his list.

The data is downloaded from the NSE website.

Co-Authored-By: Aayushmaan18 <aayushmaan18@users.noreply.github.com>