determine the support for a particular security
this code has been authored by Aayushmaan Rajora

The code is derived from calculating the resistance. That code was authored by Aprataksh Anand

Determine the support of a particular security by figuring out the maximum number of times the low of the security for a particular interval.

to do this, we divide the total price range for a particular security into several intervals and then determine how many times the low of the candles for that security have spent in that  interval.

For more accurate results, it is suggested that the candle size of the data on which the support is calculated should be as low as possible. Currently we are considering five-minute candles

Unit test: checked the support level for ashok leyland in the past few days and verified that it is correct

Co-Authored-By: Aayushmaan18 <aayushmaan18@users.noreply.github.com>