determine MACD crossover
the author of this code is Aprataksh Anand

determine the MACD values and also MACD crossover.

Instead of just pointing out the crossover, also suggest if there is a possibility of crossover.

The algorithm that is used for suggesting the possibility of crossover is pretty primitive but will be improved on at a later stage

unit test: ran this on advanced enzymes five minute stop date for the day of May 24, 2018 and visually confirmed that there are two crossovers both in the plotted chart as well as the chart on investing.com. There is a slight difference in the beginning because while the investing.com chart takes the data from the previous day for calculating its MACD at the beginning of the day, there is no such data available for the plotted chart

Co-Authored-By: aprataksh <aprataksh@users.noreply.github.com>