reorganised code for better readability and ran it for the first time successfully
reorganise the code for better readability the code that has not worked has been placed at the bottom.

Also ran with this code for the first time successfully and added the data for May 25, 2018 to the historical data for five minutes. Note that every time there is a data in addition, there is a header that is placed between the original data and the new data. This header needs to be ignored while reading the data for processing purposes

Co-Authored-By: aprataksh <aprataksh@users.noreply.github.com>