adding files downloading the date of four different duration
to make it easier to download the data for different duration, the same code has been copied with different parameters to various files.

These files will be invoked on a regular basis through another script so that the data are download process can be automated.

Unit test: tested to see if the 15 minute data is correctly downloaded.