re-factored code/add Bollinger band technical indicators
Add Bollinger band/re-factored code

Bollinger band technical indicator calculations now happen in our code instead of getting it from the ta-lib.

Also, re-factored code so that the initial steps of reading the code from the data source and sanitising itby the functions included in the utilities library

Note that the utilities library is implemented within the utilities directory that is directly underneath the project parent directory Munafa.

any code that is implementing a trading strategy should now start following the steps in this file.

The directions are being provided separately in a video

This code has been authored by Aprataksh Anand