initial code to implement open high low strategy
implement the open high low strategy.

In this strategy,  we purchase a ssecurity shortly after the opening of  the market if it meets the following criteria

the current price is near to the open price.
There has not been a large amount of either price or volume volatility since market open

The user provided configuration for this file includes

Setting for maximum price volatility - 1% price volatility, specify 0.01
setting for maximum volume volatility - tthis is in terms of the number of times the current volume  is more than the previous volume. If the current volume is twice the previous volume, specify 3

In addition, it is assumed that we will always have capital to purchase any of the securities that mmeet the aforementioned criteria. However, a single purchase  is always of a particular amount which the user can specify and is termed as the max capital for single buy

Once the securities purchased, it can be  sold when the target prices reached . If the target price is 1.5% of the purchase price, specify 0.015

if the target is not reached during the day, then the security  is assumed to be sold at 3:15 PM at the prevailing price.

The code outputs the  total purchase, total sell, profit per security listed in the master list. It also displays the number of  times the target was reachedand the number of times it had to be sold  at 3:15 PM  because the target was not reached.

When reading the historical  data, the code will skip any line that contains headers such as open. It will look at only the prices

 this code has been authored by Aayushmaan Rajora