modified the OHL strategy to look at the broader market
modified the strategy to look at the broader market also while making a decision to purchase/sell a security.

For example, if we are looking at consumer discretionary and consumer staples, we can look at the FMCG index.

Also add counters for wins and sell at the end of day. these were added as the part of the first comment but were
 inadvertently missed during the reorganisation in the second commit.

unit test: ran this with the FMCG stocks and verified the results visually for the McDonnell stock from March 13, 2018 until May 29, 2018