Add routine to check if the broader market is trending in a particular direction
this is a utility routine that will indicate whether the broader market is trending in the same direction.

The routine takes the path to the index file which contains the data of the broader market index, the direction with long or short and the date on which to evaluate the market's direction.

For long, pass one for direction. For short pass zero for direction

It returns a true or false indicating whether the market is moving in the same direction as indicated in the input parameter