fix a bug to calculate the overall cost and overall sell
 fixed that bug to calculate the overall cost and overall sell. It was being calculated by looking adding the running total cost and total sell.

This is incorrect as we are repeatedly adding earlier transactions. Instead, substituted with the total cost of the current transaction and added to the overall cost. Repeat the same procedure for the sell transaction.

Unit test: run the code and ensure that the overall cost is approximately equal to Max capital multiplied by the total number of transactions. For example if the number of buy transactions is 13 and the maximum capitalist 10,000 per transaction, then the overall cost should be equal to the 130,000. Previously it was coming to be around 180,000. This one was made for FMCG stocks from the period of March 13 until May 29 for the five-minute candle.