fix calculation of overall sell  stoploss and index
fix calculation of overall sell and purchase.
we were actually taking the price at 10 AM to calculate the overall sell.
also, the stoploss sell has to be done only if the stock has not been sold earlier due to the target having been achieved.

Also use the proper condition for checking if the index is in the same direction.

unit test: run the code and ensure that the number of purchases is equal to the total number of sell.

Also, the total overall cost should be in the vicinity of 10,000 multiplied by the total number of purchases where 10,000 as the maximum capital  per purchase