committing partial code to unblock Aayushmaan Rajora
committing partial code to display the output in CSV file, one for each stock.

In this code, we are constructing one column per data item and then hoping to concatenate all the columns together and then finally writing the entire data into one CSV file for each stock

Right now, the code is successfully creating the columns-for example, a column of all the dates that have been examined and also a column of all the buy amounts corresponding to all the transactions for one particular stock.

However, we still need to determine how to append these columns together as the zip function can take only two columns. Beyond that it is returning blank. This is happening even if we repeatedly call the zip function.