output the results in various column to CSV file
display the results in various columns in CSV file-one for each stock.

The columns either comprise of a list of conditions to decide on whether to purchase a stock and a list of conditions to decide on whether to p sell the stock. they also contain the date and the total purchase amount and the sell amount.

If a particular cell in the CSV file is not to be filled in, it will be equal to 0. for example, in the scenario where on a particular date, a purchase has not happened, then the purchase amount will be  equal to 0