Open low strategy now look set stocks preselected by daily closing higher scanner
open low strategy can now start looking at the stocks which have been selected as a result of daily closing higher scanner

the daily closing high risk and returns the list of stocks where the ) is at the end of the day is higher then the open price for a user specified consecutive number of days. 3 days is being used for testing.
So the input to the open high low strategy will be a list of stocks which have closed higher every day for the past three days from the current date that the open high low strategy is scanning.

The strategy will then simulate a purchase of the stock at 10 AM if the stock has not shown volatility in price or volume from 9:15 AM until 10 AM and the price of the stock at 10 AM is near the price at 9:15 AM.we also ensure that the stock is only purchased if the corresponding index is moving in the same direction. In this case, it should be going higher because we are purchasing the stock which means that the price at 10 AM should be greater than the price at 9:15 AM

the stock is then sold when either the user specified target profit is achieved or a user specified stoploss is met or at 3:15 PM.

this code has been co-authored by Aprataksh Anand and Aayushmaan Rajora.

Important fixes: the value of open, high, low, close can be read as strings in Python and therefore we need to convert it into float before doing any comparison or arithmetic operation.