implement end to end a daily closing higher strategy
implement the daily closing higher strategy

In this, we first select a list of stocks which are closing higher on a daily basis for a user specified number of days. The test runs have been made with a number is equal to 3.

This means that he will simulate a purchase and sale on the day number 4On the fourth day, the stocks are purchased if the price of the stock is near the closing price. The definition near is configurable and is user specified

The code has been authored by Aprataksh Anand.

Following modifications have been made to the code after the initial submission and before this commit

Display the configuration in the output
create a new function to create and initialise a new output row
create a new row if there are more than one transaction is that happen within a given day
add comments to clarify the meaning of the variable index and the multiple loops
by looking at the historical data, check for any header value which contains "close" and ignored that row

Unit test:
ran the code with nifty 50 stocks and ensured visually that the values fall within an expected spectrum