stocks can now be held and sold beyond a day
add the Ability to hold the stocks for beyond a day. Note that even though the stocks can be held for beyond a day for selling purposes, buying is restricted only on the day when there is a buy signal from the daily closing higher routine. For example, if there is a stock which is closing higher for three consecutive days, then we will consider it purchase only on the fourth day and not on the fifth or sixth or any of the following days until and unless the daily closing routine indicates that it has again closed higher on the fourth or at any other time three days in a row

the code for testing whether the index is moving in the same direction has also been added. However, since the function that implements this test has to be modified, the code is right document it out and it is assumed that it will always return true.