analyse the transactions to determine the max capital required and profit/loss
analyse the transactions file that is provided as a  output from a particular strategy and determine the maximum capital required as well as the profit/loss.

Example:
get the maximum capital available from user as input. take the output of all the buy and sell transactions from a run of daily closing higher. Then sort the file by time.
Then go through the file row by row-and for every transaction, examine if it is a buy or sell. If it is a buy transaction, debit the available balance. If it is a sell transaction creditthe available balance.

Keep a track of the minimum balance available.

If at any point of time while trying to make a purchase transaction, there is not enough capital available or in other words, the balance falls below zero if the transaction was to happen, then ignore the buy transaction and also delete the corresponding sell transaction from the list of transactions. Continue navigating the transaction file a row by row and credit the balance with any further sell transactions.

There may be some by transactions that would have happened at the end of the back test duration because there may not be any corresponding sell transactions  available as there would have been no sell signal. Those transactions will be ignored

Unit test: executed this against an output of the daily closing higher strategy and visually compared the profit and maximum capital required for two different values of maximum available capital-one is Rs. 100,000 and the second was Rs. 30,000. Also compared the profit when the available capital is Rs. 100,000 with what we got directly from the daily closing higher strategy.