sanitise the downloaded data by removing the extra headers
sanitised the downloaded data by removing the extra headers. The extra headers are included whenever we append the newly downloaded data to the historical data. At that time of writing the scripts for downloading, we did not know how to avoid addition of headers.

This becomes a problem when we are trying to process a data frame from which the headers have been removed and also look at a historical data file which includes the headers. The file is then go out of sync because of the addition of an extra header.

The code for that get bigger list function also contains a workaround to deal with files that have been shared Google Drive and have an & in the name of their corresponding ticker symbol. The files are actually stored by replacing the & with a %26 but Google Drive changes that to_.

So at the time of executing this file, it may be useful to check which version of file is there in the users directory and accordingly use the code snippet which is applicable.

This code has been authored by Aprataksh Anand

Unit test: run the file and then visually check to ensure that there is only one instance of header and that is at the beginning of the file which has been sanitised