Include the logging and printing of final transactions file
the transaction is file after looking at the maximum capital available is now printed by the ledger class.

Instead of displaying output messages on the console, start using the logging facility provided by python.

Unit test: run the code with the same data of transactions file as earlier and compared the results with the previous version. For 30,000 maximum capital, the results seem to be different as we are getting Rs. 1800 of approximate profit while we were getting Rs. 1600 of approximate profit earlier. This needs to be investigated. Apart from this, conformed that the output messages are being printed to a log file and the ledger class is printing that transactions.