refracted the code to create transactions class; add ability to use the configuration path
the transaction is class now prints the transactions to the transaction file instead of doing this in the strategy code itself. (Ideally, the transaction class should be moved into a file of its own so that it can be used by other strategies also)

Both the daily closing higher scanner as well as the daily closing higher strategy now are using the configuration files.

Here is a list of configurations files that are used

master list - contains a list of ticker symbols that are to be scanned
historical data - contains the prices at the interval mentioned in the  title of the directory-for now, we are using five-minute interval data
index_dir - contains the prices for the indexes. For now we are using five-minute interval data
- output_dir - contains the path to where the output files are stored. This is recommended to be named as Munafa_output. For every strategy, our directory can be created below this directory. At present, the daily closing higher scanner is printing its output to a separate directory for itself but ideally, it should be changed so that the scanner prints its output to our directory of the strategy that has invoked the scanner

This code has been authored by Aprataksh Anand