cleanup: more transactions class into utilities, rename classesand print all the files to a strategy specific directory
cleanup: Move the transaction class into a separate file of its own in utilities directory. This is done so that other strategies can also call this class.

Rename the main class to reflect that it is implementing daily closing high of strategy.

Ensure that transactions is now printing to the strategy specific directory. This is the same been done for logging purposes.

Unit test: run the files for the nifty 50 and ensured that the output has not changed. It should be Rs. 1602 rs profit