Cleanup: Rename strategy to scanner for daily closing higher; modify the code in the caller files
renamed the strategy daily closing higher file to scan a daily closing higher in the utilities directory. This much more accurately reflects its purpose.

Also modify the daily closing higher strategy as well as the open high low strategy files to invoke the proper methods in the scanner daily closing higher.

the Python code for transaction has again been checked in because there seem to be a problem with that in the last commit. Some of the characters in the code seem to be missing and this appears to be a github desktop issue.

Unit test: rerun the code with nifty 50 and ensured that the output remains the same as it was with the earlier run