implement daily closing higher with MACD strategy
in the daily closing hired with MACD strategy, we use a scanner to look for stocks which are closing higher on an end of day basis for a user specified number of days. We then purchase a stock when there is an MACD crossover from below-the MACD line crosses the signal line from below.

The stock is sold if the target price is met or there is a stoploss or on the end of day of the t+n days specified by the user. For example, if the user specifies three days, we will continue to look for opportunities to sell the stock purchased on the first day until the third day. We will not make any purchases on the second or third day. The purchase will will be limited to the first day

while verifying the code, it was discovered that the MACD values but not been calculated correctly. After doing the calculation by hand and then investigating this mode, it seems that we have two add an additional parameter of alpha = false in order to calculate the EMA.
More information for this parameter is can be found at https://github.com/pandas-dev/pandas/issues/7911

Given the confusion surrounding this issue, it may be useful to investigate other ways to calculate the EMA

Also, the MACD line is calculated as fast EMA minus low EMA and not as the other way round which was in the squad that was submitted the first time

This code has been authored by Aprataksh Anand