verify the RSI calculation
verify the RSI calculation by comparing them with the calculations done in Excel  at http://investexcel.net/relative-strength-index-spreadsheet/.

The data that has been used is also committed. The Excel macro file is the one that has been downloaded from the website mentioned above. The data files from Bajaj financial services were modified to include the prices from the Excel file and the calculations as a result of running the Python script were compared with that in the Excel file.

As can be inferred from the files, with RSI you 14, the initial values differ but the difference becomes progressively less and eventually becomes around 0.02 as we approach the end of the file which is about 103 rows long.

this code has been authored by Aayushmaan Rajora