at the ability to look at the index before making a decision to enter into a trade
add the ability to look at the index and decide if it is moving in the same direction before making a decision to enter into a trade

example: C we have got a buy signal at 12:15 PM on a given day. We will look at the index corresponding to that particular stock and determine the prices at market opening and at 12:15 PM. Since we are buying, we will check if the price of the index at market opening is less than the price at 12:15 PM.if so, this is a factor that will support our decision to go longer because the market is also moving in the same direction.

Unit test: run the daily closing strategy and look at the logs. Manually check the index price is printed in the log and compared with that in the raw data available.

This code has been authored by Aprataksh Anand
changes made to the court after submission:
increase the log level of messages that indicate that data for a particular date for the index that is being looked at is not available.
Remove an extra zero parameter in the stoploss function.