add the ability to download end of the data
add the ability to download end of day data.

This code differs from downloading data from shorter periods because it is using a Google Finance library. Also, it uses the config class that was written after the earlier Python scripts for downloading data.

After downloading the datafor every stock, it rearranges the columns from open high low close to close, high, low, open. This is to conform to the earlier convention of the layout of the columns that was set by the Python scripts used for downloading data.

Also, since the Google Finance library does not provide the header name for the columns which has got the date and time, we add the header ourselves.

todo: add the ability to a friend this data to the historical data as in other scripts.

Unit test: download the end of day data for nifty 500 using the script and ensure that the columns are in the right order. Also compare visually the values of Zee Entertainment limited with those available on investing.com. The values seem to differ a little bit but they are exactly the same as NSE.

This board has been authored by Aayushmaan Rajora