scanner for daily closing hired now uses end of day data
the scanner for daily closing higher now uses the end of day data instead of trying to infer the end of day values from the data available for shorter periods.

Additional parameter is added in the config file to provide the path to the historical single day data.

Unit test: ran this along with the daily closing higher script. Since the output of the scanner did not change and we had already visually checked for the scanner output in the prior runs and compare that with the daily end of day data with investing.com to confirm that the dates on which the stock was supposed to close higher was indeed correct, it can be inferred that this code also works correctly.

The sport has been authored by Aprataksh Anand