implement daily closing higher Heikin-Ashi with RSI (RSI is commented out)
implement daily closing higher with Heikin-Ashi & RSI strategy.

in this, we 1st select the list of stocks to be looked up for trading by checking which one of them is closing higher on a daily basis. This is done using the daily closing higher scanner available in the utilities directory

Once the list of stocks is selected and the output is available using the scanner.CSV file in the output directory corresponding to the strategy, we then decide on the stock to purchase using the following rules
look at the Heikin-Ashi candle (we do this on a 30 minute historical chart for execution of trading strategy) & find 2 consecutive green Heikin-Ashi candle
determine if RSI is less than 30 (we can also check if RSI has a crossover at 30)

If both the conditions are satisfied then we initiate a purchase. The sell transaction is initiated when the target is met or a stoploss is done all we have exceeded the number of days of holding the stock.

Unit tests: there were no spy signals when we combine the Heikin-Ashi along with the RSI criteria. However, if we do not consider RSI, we get by signals by using just Heikin-Ashi. On the sample trading data, for a 30 minute historical chart, the target of 1% and a stoploss of 0.5%, we are getting approximately 80% annualised profit without considering brokerage

This code has been authored by Aayushmaan Rajora