use proper index for closing values, terminate trading at 3 PM, including brokerage
this commit includes the following fixes/enhancements

use the proper index for closing values when attempting to sell a stock. So far, we were using the highest value.

The end of file test should include the length of the file - 1 instead of just the length of the file

When we encounter a case where we cannot find the date identified by the daily closing higher scanner in the historical data, we print an error message as data is corrupted and moving to the next stock and continue the process. This issue can result when the date I for their daily closing higher as later dates, for example June 16, 2018 while the data for historical five-minute is limited to dates prior to June 16, 2018

Use brokerage charges to determine the actual profit. https://zerodha.com/charges is taken as an example

we do not make any purchase or sell after 3 PM. So far, we had been including transactions made at 3:30 PM which is not possible because the market closes and we get the trade signals at the same time. We can possibly do this a few minutes before 3:30 PM but if we are using the historical 30 minute data, the earliest time we would have is 3 PM and therefore for now, we are limiting trading activity after 3 PM. This can be modified later on depending on the granularity of the data that we have

also display the number of days that the daily closing higher scanner uses in configuration as well as displayed the profit after accounting for the brokerage charges and the percentage wins and the intraday sales.
unit test: run the daily closing higher test again and manually calculate the brokerage fee and compare with the results. Also compare with the results that we adopt end earlier with similar target and stoploss parameter and ensured that they are of the same order of magnitude. In this case, we were getting around Rs. 6000 a profit and now we're getting Rs. 4100 on a similar capital.

This code has been partly authored by Aprataksh Anand