the output of the scanner is now dependent on the user specified number of days for which to scan
since this scanner was initially quoted for three days, it would output the rates for three days no matter what was the user specified number of days to scan. For example, even if it is scanning for two days, it wouldfour days starting from the t-2 to t+1 day, with t as the current day it is scanning.

with the changes, the scanner will print the number of days for which it has to scan plus the next day on which we are supposed to trade. The other strategies are supposed to pick up the last column and use the dates they are to perform a buy or sell operation.

Due to the difference in output, there was a major bug for n=2, where we would pick of the third column which would contain the second day when the stock closed higher on a daily basis and we would purchase and sell on that same day. Since we have already proven that the stock closed higher on the second day, purchasing and selling the stock on that day would greatly improve the resultsleading to numerous false positives.

Unit test: run the code and manually verify a few entries.

This code has been implemented by Aprataksh Anand