use proper index for closing values, terminate trading at 3 PM etc
this, it is essentially a port of https://github.com/Aprataksh/Munafa/commit/e4187474e870241023d3a453ed6b5327192f1d9b.

it is a copy of the daily closing higher with RSI. A separate file was made because with RSI of 30 minutes historical data and to green candles of Heikin-Ashi, we were not getting a buy signal. We are getting a buy signal only to green candles of Heikin-Ashi on historical 30 minute chart and therefore we thought to create a separate file for this.

Apart from this, there are also changes in the terminating condition when we reach the end of the file. There was an error when we reach the end of the file but the closing time is 3 PM. If the end of file is 3:30 PM, then we will not be able to sell the stock because we are specifically looking for a closing time of 3 PM or any other user specified time when we are at the end of the file. But usually at the end of the file we are at 3:30 PM. Therefore we end up not selling the stock.

This code has been authored by Aayushmaan Rajora

Unit test: run the test with historical 30 minute data