a more robust way to download intraday data
this code is a more robust way to download intraday data. As compared to the earlier code, it will do trivial validation like counting the number of file sand comparing it with the number ofticker symbols in the master list.

It also appends the file to the historical data without the extra headers.

the same code can be used with different config to download data for different time intervals. The time intervals is in seconds and need to be modified in the code

Unit test: download the 30 minute data and ensure that it has been properly appended. The last 30 minute data download was till June 11, 2018 and visually check to see it now couple of files that the data in June 12, 2018 has been properly appended