add the ability to use a trailing stop loss
add the ability to use a trailing stop loss. The trailing stoploss is explained with an example.. Be used to additional parameters. The first additional parameter is the initial target. The second parameter is called as the trailing stoploss.

Let's assume the initial target to be 1% and the final target to be 3%let's assume the initial stoploss to be -0.5%. Then we meet the initial target of 1%, we move our initial stop lossfrom -0.5% to the value of trailing stoploss. Let us assume that to be 0.5%.

Now if the price goes down from a gain of 1% to a gain of 0.5%, we will sell. If the price goes up from 1% to 1.1%, we will increase our stoploss from 0.5% to 0.6%.  In other words, the stoploss is going to start trailing the price.

This code has been authored by Aprataksh Anand

Unit test: execute the strategy and check the prices of the first few stocks which have trailing stoploss triggered. Match the prices by hand and ensured that the trailing stoploss calculations are accurate