Heikin-Ashi code now get directly invoked from a strategy
modify the Heikin-Ashi code so that it can get directly invoked from a strategy. Earlier Heikin-Ashi code had to be run separately and then the output path had to be provided to a strategy.

Also modify the daily closing higher with Heikin-Ashi to do more to any references to RSI. Earlier the code was partially commented out and it still required the output from RSI to be present.

This code has been authored by Aayushmaan Rajora.

Unit test: run the modified Heikin-Ashi code of the daily closing higher with RSI strategy and compare the results with the previous run with the same parameters. Ensure that the results are exactly the same