the RSI function can now be invoked by strategy
the RSI function can now be invoked by a strategy code. Specifically the first use is for it to be invoked to the daily closing higher with RSI strategy.

Also, make the RSI files that are a part of output read-only so that they are not accidentally modified.

This code has been authored by Aprataksh Anand

Unit test: none. The changes were trivial and so far we have not been able to find any particular use of RSI and therefore we are not running the strategy for now