featured for trailing stop loss included; rename file
the feature for trailing stoploss has now been included. Also made modifications to the termination condition when we reach the end of file.

Also renamed the file to not have any_'s.

Unit test: no test has been done as yet. This is an integration of the code which was first committed for daily closing higher and can be found at

https://github.com/Aprataksh/Munafa/commit/624be053d9baff2458a5c779fa39c681835881d0