add the ability to short a stock
add the ability to short a stock. To do this, we essentially multiply the calculations were to be a determining the change in prices by -1.

We also modify  the condition in trailing stoploss to look at less than instead of greater than when we compare the price with the purchase price

Depending on the flag for the long or short position, we look at either to green Heikin-Ashi candles so long position or to read Heikin-Ashi candles for taking a short position.

After the submission from Aayushmaan Rajora, added the following:
modified the condition for checking if the initial target has been met by comparing if the current price is less than the purchase price instead of comparing if it is greater than the purchase price.
Also modified the display statements so that the display is spread over a few lines instead of one continuous single line.
Also started displaying the configuration for the trailing stoploss in the output..
The flag whether to use a trailing stoploss is now added in the constructor

Unit test:
the regression test by checking if the results for taking up the long position are exactly the same as they were before making these changes. The results which matched had the following conditions. There is no trailing stoploss, the final target is 1.5%, the stoploss is at -0.8%, the maximum days of holding is 3, the maximum number of days of daily closing higher is 2, we are disregarding the index and only looking for consecutive 2 Heikin-Ashi candle which are green. With this, the number of trades is 245 and the loss after the run of this strategy is approximately Rs. -3000.

Also did a testing and manually verified that it works for one of the stocks with a trailing stoploss when we take up a short position.

This quote has been submitted by Aayushmaan Rajora