while appended to final date, ignore any ticker symbol which has no corresponding downloaded data
ignored any ticker symbol which does not have downloaded data associated with it while appending the downloaded data with the earlier data.

It may happen that there is no data associated with a particular ticker symbol possibly owing to the lack of transactions because of no trading happening.

We need to ignore such data but it will be useful to have a display statement associated so as to alert the user.

Unit test: download data for 1 hour and showed it goes through for the files which were not having any data associated with them

This code has been submitted by Aprataksh Anand