reorganise the code into base class of buy and sell
in order to make the code reusable and reduce the modifications needed whenever we change the buy and sellfunctions, a base class has been created for buy and sell and at the strategy classes will now be inherited from that.

There is no other material functional code change in this file.

Unit test: do a regression test to see if the code produces the same results as earlier. The regression test was done with target price as 1.5%, the stoploss is -0.8%, there is no trailing stoploss, we are assuming a long position, the number of days to hold the stock is 3, the total number of days for daily closing higher is 2.

This code has been submitted by Aayushmaan Rajora